device mapper / lvm2 support for mkinitrd (by sumit.bose@suse.de).
in detail:

Because of having / on LVM I've patched mkinitrd from mkinitrd-1.0-102
a bit, the diff is attached.

The first change is needed, because when e.g. vga=0xabc is scanned the
old
version finds only the 0 of 0x.

The second block put the LVM2 tools into the initrd.

The next block activated the lvm commands when a DM root partition was
found
and ramfs is used instead of shm.

In the next block I changed the dection of the major number of the root
partition so that symlinks are followed.

Finally dm-mod is added to initrd when a root partition on DM was found.