Unknown modules treated as warning, module exclusion
Usually if using a stock SuSE kernel everything is built modular.
If you build your own kernel, you can define several parts of it as
being built-in into the kernel, so there is no module for this
particular feature. This patch changes this case from an error
to a warning message, so the user can decide on his own if the
problem is fatal and the initrd is built anyways.

Additionally when a kernel module preceding a minus sign is added to
the module list, this is interpreted as "do not include this module",
so one can manually exclude modules from the build process.

Signed-off-by: Alexander Graf <agraf@suse.de>