Rework multipath device detection
For (or any fc HBA, really) we cannot be sure that all devices
are discovered after the module is loaded. So while we're
already waiting for 'normal' block devices, we just run
multipath once and hope that all devices are there.
Not good.
So better rework the udev_wait_for_devices loop to check the
major number of the found device and call multipath if it's
not a device-mapper device. And only terminate the loop
if the device is in fact a device-mapper device.

References: 369468
Signed-off-by: Hannes Reinecke <hare@suse.de>