Use the same password for each LUKS device
In SP2 support for encrypted LUKS root partitions was added. When encrypting
partitions a script in the initrd asks the user for the password. In SP2 the
password is asked for each encrypted partition. In the patsch supplied by DFS
the user is asked only once for the password if the identical password was used
for all partitions. The script provides by Novell mimics this behavior but the
variable retva which is set  in the script seems to have the wrong/no value.
This causes the repeated request for a password. In our case the Laptop users
are asked 5 times for the same password.

References: 396346
Signed-off-by: Hannes Reinecke <hare@suse.de>