Always rescan for LVM when using mulitpath
We have a race condition when running LVM on top of multipathing.
As the device scanning is done in the background it might well
be that devices appear after multipath has processed all available
devices but before LVM starts. So we have to disable LVM before
multipath is called and also do a final scan after the device
has been discovered.

Signed-off-by: Hannes Reinecke <hare@suse.de>