Always include all ATA and SCSI devices
This patch changes the ATA and SCSI default behavior. Usually
we copied over only the device driver for the devicees we do
detect.

Including all drivers increases the size of the initrd by
1.5 MB. As a benefit we get more flexibilty on which computers
the system will run, which makes a lot of things easier.

The patch also introduces a new option: -S
This option enforces the old behavior.

Signed-off-by: Alexander Graf <agraf@suse.de>
Signed-off-by: Hannes Reinecke <hare@suse.de>