Include all libata drivers only if libata is used for the root device
Due to a recent change, the default was to include all block device drivers.
This default is improper, as it increases the bootup times for machines
that have multiple controllers and only need to initialize one to get into
the root system.

This patch changes the default behavior to include all libata drivers when
libata was found, but does not include other superfluous modules.

Signed-off-by: Alexander Graf <agraf@suse.de>
Signed-off-by: Hannes Reinecke <hare@suse.de>