Resolve bridge network devices when auto-detecting the network interfaces
For the case that a bridge contains only one network device as usual when
using Xen or some other visualization environment (KVM, VirtualBox at least),
resolve the bridge name. This means when the bridge br0 contains only eth0,
we use eth0 as automatic network interface.

The patch is important to be able to use mkinitrd ... -D default to build a
kdump initrd (calling mkinitrd(8) from mkdumprd(8)).

Signed-off-by: Bernhard Walle <bwalle@suse.de>