Add handling for unsupported kernels
mkinitrd doesn't add kernel modules which are unsupported. However, that does
not make sense when you have a non-SUSE kernels because it prevents users from
using a non-SUSE kernel on SLES at all without changing
/etc/modprobe.d/unsupported-modules.

This patch checks if the 'ext3.ko' (because it's always present) module
is unsupported. If it is, then we assume that the kernel is unsupported as a
whole. Then we add that kernel modules to initrd.

Because the patch in module-init-tools checks if we're running on a unsupported
kernel and loads the kernel modules then, that initrd is actually usable.

Signed-off-by: Bernhard Walle <bwalle@suse.de>