Add 'SUSE INITRD foo REQUIRES bar' comment parsing
To express dependencies of modules that use install lines in /etc/modprobe.conf
and friends we add a new comment syntax

    # SUSE INITRD foo REQUIRES bar

in /etc/modprobe.conf, /etc/modprobe.conf.local and /etc/modprobe.conf.d/*
files. That allows us for example to add

    # SUSE INITRD: uhci-hcd REQUIRES ehci-hcd
    install uhci-hcd /sbin/modprobe ehci-hcd; /sbin/modprobe --ignore-install uhci-hcd
    # SUSE INITRD: ohci-hcd REQUIRES ehci-hcd
    install ohci-hcd /sbin/modprobe ehci-hcd; /sbin/modprobe --ignore-install ohci-hcd

in /etc/modprobe.conf. That makes that install lines actually working because
then mkinitrd includes the ehci-hcd module if uhci-hcd or ohci-hcd is required.

We need to replace that for Code 11++ (openSUSE 11.2 and later).

References: bnc #449853

Signed-off-by: Bernhard Walle <bwalle@suse.de>