Include virtio_pci if necessary
When using virtio_net as network device, we also need virtio_pci (and its
dependency virtio_ring) in the initrd. The dependency isn't expressed in
udev rules or install-lines (/etc/modprobe.conf*). The virtio_pci driver is just
loaded by the modalias

    pci:v00001AF4d*sv*sd*bc*sc*i*

in the module. Therefore we have to hardcode that dependency, at least for now.

References: bnc #456439

Signed-off-by: Bernhard Walle <bwalle@suse.de>