Fix libata detection
Now by default all libata drivers are included if a disk is detected as libata.
One could use the -S option to disable that behaviour.

That was already implemented, but didn't work. The problem was that a global
variable called block_uses_libata was set, but the function that sets that
variable was called in a subshell like

    bla=$(handle_scsi ..)

To solve the problem we now don't use a subshell but use indirect variable
references to get the result (see http://tldp.org/LDP/abs/html/complexfunct.html
and the example 23-5 to get a documentation how that indirect variable
references are used in bash).

Signed-off-by: Bernhard Walle <bwalle@suse.de>