Fix handling of minor/major device numbers for minor device numbers > 255
New kernels may use minor device numbers larger than 255 for storage. One can
enable that for debugging with CONFIG_DEBUG_BLOCK_EXT_DEVT set in the kernel
configuration.

That patch changes the handling of minor/major device numbers in a way
that the first 20 bits are used for the minor number and the rest for the
major number. That's what the kernel currently does, so we're on the safe
side.

Signed-off-by: Bernhard Walle <bwalle@suse.de>
Cherry-picked from 9d62c04e231c35087e1a1e98630205da8233737f
Signed-off-by: Michal Marek