Use vmlinuz on x86
On x86 (i386/x86_64) the shell glob to find the default kernel was
'vmlinu.', so both 'vmlinux' and 'vmlinuz' did match. Because 'x' is
before 'z', if we have the same file as 'vmlinux' and 'vmlinuz', the
match was 'vmlinux'.

However, we use the bzImage (vmlinuz) for booting on x86, so it doesn't
make sense to match for vmlinux (the ELF image which cannot be used to
boot the system with GRUB/LILO).

Signed-off-by: Bernhard Walle <bwalle@suse.de>