Fix detection of compiled-in modules
The detection of compiled-in modules didn't work properly in all cases.
The 'foo REQUIRES bla' syntax in the modprobe configuration files
were not catched.

That compiled-in modules checking is needed for PPC64 because their
kernel configuration contains compiled-in stuff (mostly USB) that other
architectures have as module in the default SUSE configuration. However,
it may also be useful on other architectures for self-compiled kernels.

References: bnc #449123
Signed-off-by: Bernhard Walle <bwalle@suse.de>