Remote the -t parameter and use /tmp as user
Remove the -t (temporary directory) parameter because we need also a
temporary directory for parameter parsing. So it doesn't make sense.

Instead, use the TMPDIR environment. If unset, default to /dev/shm
when writable (as root) and to /tmp otherwise (as user).

This fixes /sbin/mkinitrd -h as user.

References: bnc #471953
Signed-off-by: Bernhard Walle <bwalle@suse.de>