MD and multipath doesn't work together
After update to SLES10 SP2 from SLES10 SP1 the system root is not found during
boot. The system is running on a softraid device consisting of 2 local disks.
After debugging of the init script from the initrd he found that the following
code is at failure:

            if [ -e $root ]; then
                devn=$(devnumber $root)
                major=$(devmajor $devn)
                if [ -z "$dm_major" ] || [ "$major" == "$dm_major" ] ; then
                    retval=0
                    echo " ok"
                    break;
                else
                    echo -n "!"
                    multipath -v0
                   /sbin/udevsettle --timeout=$udev_timeout
                fi
            fi

The comparison between $major and $dm_major seems to be problematic.
Fixed by checking if the major number is a md major number and
just return if true.

References: 449702

Signed-off-by: Hannes Reinecke <hare@suse.de>