lfpc: wait 15 seconds for linkup
During initialisation the link state will be place in 'link down',
leaving us unable to identify whether a link down state is due to a link
not detected or due to the initialisation sequence not finished.
According to the lpfc sources 15 seconds should be enough to wait for a
link up. So when no link was detected after 15 seconds we can assume the
driver hasn't been able to establish a link.

References: bnc#635869