[netconsole]: Add deadline to initial ping command.
So far 'single shot' ping was sent to the destination address using
the '-c1' option. This timed out after two RTT (default for ping).
If the network interface was not quite ready when ping tried to send
the packet it was waiting for a reply that never came - with a timeout
that's quite high for a local network.
Giving ping a deadline, it will send packages until the deadline is
exceeded or until the number of reply packages as specified with the
'-c' option have been received.
This results in a more reliable operation in this context and a shorter
timeout if the other end doesn't respond.