Switch plymouth and/or splash off before escaping to emergency shell
As well as use sulogin to have a emergency shell on all consoles and
to enable a password check.  For this check the initrd should be set
to be readable only by root to avoid that users can to spy on the
confidential data stored within the initrd.

References: bnc#809812
Signed-off-by: Werner Fink <werner@suse.de>
Signed-off-by: Michal Marek <mmarek@suse.cz>