make initrd content even more safe (take 2)
If the target file exists already, cp doesn't change permissions.

The previously attempted "chmod 0600" is an incomplete fix,
as someone might have an open file descriptor to the file.
Removing the old file before copying would work, but leave the
user without a working initrd if the copy fails.

Thus copy to a new mktemp file on the same file system as the target
(mktemp always creates files with mode 0600), which can then be
atomically moved to the final path.