make initrd content even more safe (take 3)
The previous attempt tried to avoid to leave customers without
an initrd by creating the new image under /boot first, and only
overwriting the existing image after that succeeded. But that has the
disadvantage to need considerable extra disk space under /boot,
which implies a regression risk for systems with small /boot
partitions. That risk is higher than the copy process failing
for other reasons. Thus we simply remove the existing initrd image
now and rely on cp using the permissions of the source file for
the target.